+++ tests/state_spec.lua
@@ -10,15 +10,6 @@ local function validate(opts)
     return state.validate()
 end
 
----@param opts render.md.UserBufferConfig
----@return string[]
-local function validate_override(opts)
-    return validate({ overrides = { buftype = { nofile = opts } } })
-end
-
-local prefix = 'render-markdown'
-local override_prefix = prefix .. '.overrides.buftype.nofile'
-
 describe('state', function()
     it('valid', function()
         eq(0, #validate())
@@ -28,76 +19,114 @@ describe('state', function()
         vim.bo.buftype = 'nofile'
         eq(false, state.get(0).sign.enabled)
 
-        eq(0, #validate({ callout = { note = { raw = 'value' } } }))
-
-        eq(0, #validate({ callout = { new = { raw = 'value', rendered = 'value', highlight = 'value' } } }))
-
-        eq(0, #validate_override({
-            enabled = false,
-            max_file_size = 0,
-            render_modes = {},
-            anti_conceal = {},
-            heading = {},
-            code = {},
-            dash = {},
-            bullet = {},
-            checkbox = { unchecked = {}, checked = {} },
-            quote = {},
-            pipe_table = {},
-            callout = {},
-            link = {},
-            sign = {},
-            win_options = {},
+        eq(0, #validate({
+            callout = {
+                note = { raw = 'value' },
+                new = { raw = 'value', rendered = 'value', highlight = 'value' },
+            },
+            overrides = {
+                buftype = {
+                    nofile = {
+                        enabled = false,
+                        max_file_size = 0,
+                        render_modes = {},
+                        anti_conceal = {},
+                        heading = {},
+                        code = {},
+                        dash = {},
+                        bullet = {},
+                        checkbox = { unchecked = {}, checked = {} },
+                        quote = {},
+                        pipe_table = {},
+                        callout = {},
+                        link = {},
+                        sign = {},
+                        win_options = {},
+                    },
+                },
+            },
         }))
     end)
 
     it('extra', function()
-        eq({ prefix .. '.additional: is not a valid key' }, validate({ additional = true }))
-
         eq(
-            { prefix .. '.callout.note.additional: is not a valid key' },
-            validate({ callout = { note = { additional = true } } })
+            {
+                'render-markdown.additional: is not a valid key',
+                'render-markdown.anti_conceal.ignore.additional: is not a valid key',
+                'render-markdown.callout.note.additional: is not a valid key',
+                'render-markdown.checkbox.checked.additional: is not a valid key',
+                'render-markdown.latex.additional: is not a valid key',
+                'render-markdown.overrides.additional: is not a valid key',
+                'render-markdown.overrides.buftype.nofile.additional: is not a valid key',
+            },
+            validate({
+                additional = true,
+                anti_conceal = { ignore = { additional = true } },
+                callout = { note = { additional = true } },
+                checkbox = { checked = { additional = true } },
+                latex = { additional = true },
+                overrides = { additional = true, buftype = { nofile = { additional = true } } },
+            })
         )
-
-        eq({ prefix .. '.latex.additional: is not a valid key' }, validate({ latex = { additional = true } }))
     end)
 
     it('type', function()
+        ---@diagnostic disable: assign-type-mismatch
         eq(
-            { prefix .. '.enabled: expected boolean, got string' },
-            ---@diagnostic disable-next-line: assign-type-mismatch
-            validate({ enabled = 'invalid' })
+            {
+                'render-markdown.anti_conceal.ignore.sign: expected string list or type boolean or nil, got string',
+                'render-markdown.callout.a.highlight: expected type string, got nil',
+                'render-markdown.callout.b: expected type table, got string',
+                'render-markdown.checkbox.checked: expected type table, got boolean',
+                'render-markdown.checkbox.unchecked.icon: expected type string, got boolean',
+                'render-markdown.debounce: expected type number, got table',
+                'render-markdown.enabled: expected type boolean, got string',
+                'render-markdown.heading.enabled: expected type boolean, got string',
+                'render-markdown.log_level: expected one of { "debug", "info", "error" }, got string',
+                'render-markdown.log_runtime: expected type boolean, got string',
+                'render-markdown.max_file_size: expected type number, got boolean',
+                'render-markdown.overrides.buftype.nofile.sign.highlight: expected type string or nil, got boolean',
+                'render-markdown.overrides.buftype.other: expected type table, got boolean',
+                'render-markdown.overrides.filetype: expected type table, got string',
+                'render-markdown.padding.highlight: expected type string, got boolean',
+                'render-markdown.preset: expected one of { "none", "lazy", "obsidian" }, got string',
+                'render-markdown.render_modes: expected string list or type boolean, got string',
+            },
+            validate({
+                anti_conceal = { ignore = { sign = 'invalid' } },
+                callout = {
+                    a = { raw = 'value', rendered = 'value' },
+                    b = 'invalid',
+                },
+                checkbox = { checked = false, unchecked = { icon = false } },
+                debounce = {},
+                enabled = 'invalid',
+                heading = { enabled = 'invalid' },
+                log_level = 'invalid',
+                log_runtime = 'invalid',
+                max_file_size = true,
+                overrides = {
+                    buftype = {
+                        nofile = { sign = { highlight = false } },
+                        other = false,
+                    },
+                    filetype = 'invalid',
+                },
+                padding = { highlight = true },
+                preset = 'invalid',
+                render_modes = 'invalid',
+            })
         )
 
         eq(
-            { prefix .. '.log_level: expected one of { "debug", "info", "error" }, got invalid' },
-            ---@diagnostic disable-next-line: assign-type-mismatch
-            validate({ log_level = 'invalid' })
-        )
-
-        eq(
-            { prefix .. '.render_modes: expected string list or type { "boolean" }, got invalid' },
-            ---@diagnostic disable-next-line: assign-type-mismatch
-            validate({ render_modes = 'invalid' })
-        )
-
-        local int_render_modes = { 1, 2 }
-        eq({
-            prefix
-                .. '.render_modes: expected string list or type { "boolean" }, got '
-                .. tostring(int_render_modes)
-                .. '. Info: [1] is number',
-        }, validate({ render_modes = int_render_modes }))
-
-        eq(
-            { prefix .. '.callout.new.highlight: expected string, got nil' },
-            validate({ callout = { new = { raw = 'value', rendered = 'value' } } })
-        )
-
-        eq(
-            { override_prefix .. '.sign.highlight: expected string, got boolean' },
-            ---@diagnostic disable-next-line: assign-type-mismatch
-            validate_override({ sign = { highlight = false } })
+            {
+                'render-markdown.checkbox: expected type table, got string',
+                'render-markdown.render_modes: expected string list or type boolean, got table, info: [1] is number',
+            },
+            validate({
+                checkbox = 'invalid',
+                render_modes = { 1, 2 },
+            })
         )
     end)
 end)