+++ lua/render-markdown/lib/icons.lua
@@ -7,11 +7,14 @@ local M = {}
 ---@param language string
 ---@return string?, string?
 function M.get(language)
+    -- Handle input possibly being an extension rather than a language name
+    local file_type = vim.filetype.match({ filename = 'a.' .. language }) or language
+
     if has_mini_icons then
         ---@diagnostic disable-next-line: return-type-mismatch
-        return mini_icons.get('filetype', language)
+        return mini_icons.get('filetype', file_type)
     elseif has_devicons then
-        return devicons.get_icon_by_filetype(language)
+        return devicons.get_icon_by_filetype(file_type)
     else
         return nil, nil
     end